+++ CHANGELOG.md
@@ -37,6 +37,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `ember-cli-mirage` - the 70% that seems possible to express in typescript
 - Engines:
     - `ember-osf-registries` - moved/upgraded into the registries engine
+- Handbook:
+    - `tags-widget` - added to the handbook
 
 ### Changed
 - Models: